backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
deep-translator==1.11.4
defusedxml==0.8.0rc2
Django==5.1.6
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
ecdsa==0.19.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
